neural-style-transfer/requirements.txt
numpy
opencv-python-headless
Pillow
streamlit
tensorflow
tensorflow-hub
